Open a conversation in a right-hand drawer, not a new page
The canvas's `对话详情` artboard was redrawn this round as `对话详情 · PC
（右侧抽屉）`, and its note sets four rules. Reading a conversation no longer
replaces the screen: the panel slides in from the right at 880px and the list
stays where it was, narrowed to a 328px rail.

No scrim. The list stays lit and clickable, so moving to the next conversation
is one click rather than back-then-forward — which is the whole reason for a
drawer, and what a scrim would take away.

The drawer header carries the neighbours: previous and next with J and K, the
position in the list, and a close that Escape repeats.

The URL still becomes `/chats/:id`, so links, refreshes and the back button
behave as before, and opening one directly lands on the list with the drawer
already open. `/chats` and `/chats/:id` therefore resolve to the same screen,
and the detail page becomes a variant of one panel rather than its own page.

Below 1180px there is no room for both, and the conversation takes the screen
on its own. That decision lives in the page rather than the route, so dragging
a window across the threshold keeps the conversation instead of dropping the
reader back on the list.

The note leaves one question open — whether closing with an unsaved inline
edit should block or save. Closing commits an open title edit and asks
nothing, per the leaning; the markdown and list editors keep their explicit
save, since committing a half-written body silently is worse than asking.

Verified against the new artboard at 1440, in both themes, plus the fallback
at 1100 and 390 and a live resize across the threshold; `npm run build` clean.
